refactor: extract bulk actions toolbar for data-table
